requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
cryptography==47.0.0
decorator==5.2.1
distro==1.9.0
executing==2.2.1
frozenlist==1.8.0
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.47
google==3.0.0
google-auth==2.49.2
google-genai==1.73.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
ipython==8.39.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.1
narwhals==2.20.0
numpy==2.4.4
openai==0.27.10
packaging==26.2
pandas==3.0.2
parso==0.8.6
pillow==12.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
pure_eval==0.2.3
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
streamlit==1.56.0
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.3.3
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
